refactor: restructure service packages
