+++ webtrit_callkeep_android/android/src/main/kotlin/com/webtrit/callkeep/ConnectionsApi.kt
@@ -2,8 +2,8 @@ package com.webtrit.callkeep
 
 import com.webtrit.callkeep.common.toSignalingStatus
 import com.webtrit.callkeep.models.toPConnection
-import com.webtrit.callkeep.services.dispatchers.SignalingStatusDispatcher
-import com.webtrit.callkeep.services.telecom.connection.PhoneConnectionService
+import com.webtrit.callkeep.dispatchers.SignalingStatusDispatcher
+import com.webtrit.callkeep.services.connection.PhoneConnectionService
 
 class ConnectionsApi() : PHostConnectionsApi {
     override fun getConnection(
