+++ webtrit_callkeep_android/android/src/main/kotlin/com/webtrit/callkeep/managers/NotificationManager.kt
@@ -6,9 +6,9 @@ import androidx.core.app.NotificationManagerCompat
 import com.webtrit.callkeep.common.ContextHolder.context
 import com.webtrit.callkeep.models.CallMetadata
 import com.webtrit.callkeep.notifications.MissedCallNotificationBuilder
-import com.webtrit.callkeep.services.ActiveCallService
-import com.webtrit.callkeep.services.incomming_call.IncomingCallRelease
-import com.webtrit.callkeep.services.incomming_call.IncomingCallService
+import com.webtrit.callkeep.services.active_call.ActiveCallService
+import com.webtrit.callkeep.services.incoming_call.IncomingCallRelease
+import com.webtrit.callkeep.services.incoming_call.IncomingCallService
 import io.flutter.Log
 
 class NotificationManager() {
