+++ webtrit_callkeep_android/android/src/main/kotlin/com/webtrit/callkeep/notifications/ActiveCallNotificationBuilder.kt
@@ -10,7 +10,7 @@ import com.webtrit.callkeep.common.ContextHolder.context
 import com.webtrit.callkeep.models.CallMetadata
 import com.webtrit.callkeep.models.NotificationAction
 import com.webtrit.callkeep.managers.NotificationChannelManager.NOTIFICATION_ACTIVE_CALL_CHANNEL_ID
-import com.webtrit.callkeep.services.ActiveCallService
+import com.webtrit.callkeep.services.active_call.ActiveCallService
 
 class ActiveCallNotificationBuilder() : NotificationBuilder() {
     private var callsMetaData = ArrayList<CallMetadata>()
